requirements.txt
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
pdfplumber>=0.10.0
Jinja2>=3.1.0
RapidFuzz>=3.5.0
premailer>=3.10.0
python-dotenv>=1.0.0
